Fixed bugs - update theme selector to change background theme, apply background theme to preview and live page

+++ app/_components/common/input/workspace-hero-uploader.jsx
@@ -40,13 +40,15 @@ const HeroWorkspaceUploader = ({ uri }) => {
     const { permanentSignedUrl, response } = await uploadFile(file);
 
     try {
-      deleteFile({
-        file: previousImage,
-        BUCKET,
-        ACCESSKEY,
-        SECRETKEY,
-        ENDPOINT,
-      });
+      if (previousImage) {
+        deleteFile({
+          file: previousImage,
+          BUCKET,
+          ACCESSKEY,
+          SECRETKEY,
+          ENDPOINT,
+        });
+      }
       updateHero({
         ...hero,
         extraAttributes: {
